% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropic_verify.R
\name{anthropic_verify}
\alias{anthropic_verify}
\title{anthropic_verify}
\usage{
anthropic_verify(...)
}
\arguments{
\item{...}{}
}
\value{
True if the parameters are valid, False otherwise.
}
\description{
anthropic_verify
}
