% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude.R
\name{claude}
\alias{claude}
\title{claude}
\usage{
claude(
  source,
  input,
  output = "output",
  prompt,
  return_invisible = FALSE,
  iterations = 1,
  model = "claude-3-haiku-20240307",
  anthropic_version = "2023-06-01",
  temperature = 1,
  top_p = NULL,
  max_tokens = 4096,
  top_k = NULL,
  anthropic_api_key = Sys.getenv("ANTHROPIC_API_KEY"),
  progress = TRUE
)
}
\arguments{
\item{source}{required; A source dataframe}

\item{input}{required; A column name in the source dataframe}

\item{output}{optional; A string of a column name (Or a vector of strings) to be created in the source dataframe storing the output of the models. Defaults to \code{output}.}

\item{prompt}{required; A string (Or vector of Strings for handling multiple operations at the same time) of a system message to be sent to the AI model.}

\item{return_invisible}{optional; A boolean to return just the output (\code{TRUE}) or an llm object containing model metadata (\code{FALSE}). Defaults to \code{FALSE}.}

\item{iterations}{DEV ONLY. SUPPORT COMING SOON. Number of completions to generate. Integer. Defaults to \code{1}.}

\item{model}{required; a length one character vector.}

\item{anthropic_version}{optional; defaults to \code{2023-06-01}; a length one character vector. Specifies the version of the Anthropic's models.}

\item{temperature}{optional; defaults to \code{1}; a length one numeric vector with the value between \code{0} (More analytical) and \code{1} (More creative).}

\item{top_p}{optional; a length one numeric vector with the value between \code{0} and \code{1}. Only specify a temperature or a top_p, never both. Not recommended, for most use cases use temperature instead.}

\item{max_tokens}{optional; defaults to \verb{(4096 - prompt tokens)}; a length one numeric vector with the integer value greater than \code{0}.}

\item{top_k}{optional; a length one numeric vector with the integer value greater than \code{0}. Only sample from the top_k options for each subsequent token. Not recommended, for most use cases use temperature instead.}

\item{anthropic_api_key}{required; defaults to \code{Sys.getenv("ANTHROPIC_API_KEY")} (i.e., the value is retrieved from the \code{.Renviron} file); a length one character vector. Specifies Anthropic API key.}

\item{progress}{optional; a length one logical vector. Defaults to \code{TRUE}. Determines whether to show a progress bar in the console.}
}
\value{
A dataframe with the output column(s) created
}
\description{
claude
}
