% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBAI-package.R
\docType{package}
\name{DBAI-package}
\alias{DBAI}
\alias{DBAI-package}
\title{DBAI: OpenAI Tools for R Datasets}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Adam-Hubbs/DBAI}
  \item Report bugs at \url{https://github.com/Adam-Hubbs/DBAI/issues}
}

}
\author{
\strong{Maintainer}: Adam Hubbs \email{amhubbs17@gmail.com}

}
\keyword{internal}
