% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.character.R
\name{gemini.character}
\alias{gemini.character}
\title{gemini}
\usage{
\method{gemini}{character}(
  source,
  output = "output",
  prompt = "",
  model = "gemini-1.5-flash",
  iterations = 1,
  repair = FALSE,
  progress = TRUE,
  temperature = 1,
  top_p = NULL,
  top_k = NULL,
  max_tokens = 4096,
  google_api_key = Sys.getenv("GOOGLE_API_KEY"),
  call = rlang::caller_env(),
  parentInfo = NULL
)
}
\description{
gemini
}
