% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini}
\alias{gemini}
\title{gemini}
\usage{
gemini(
  source,
  input,
  output = "output",
  prompt,
  return_invisible = FALSE,
  iterations = 1,
  model = "gpt-3.5-turbo",
  temperature = 1,
  top_p = 1,
  n = 1,
  max_tokens = 4096,
  presence_penalty = 0,
  frequency_penalty = 0,
  progress = TRUE
)
}
\arguments{
\item{source}{required; A source dataframe}

\item{input}{required; A column name in the source dataframe}

\item{output}{optional; A string of a column name (Or a vector of strings) to be created in the source dataframe storing the output of the models. Defaults to \code{output}.}

\item{prompt}{required; A string (Or vector of Strings for handling multiple operations at the same time) of a system message to be sent to the AI model.}

\item{return_invisible}{optional; A boolean to return just the output (\code{TRUE}) or an llm object containing model metadata (\code{FALSE}). Defaults to \code{FALSE}.}

\item{iterations}{DEV ONLY. SUPPORT COMING SOON. Number of completions to generate. Integer. Defaults to \code{1}.}

\item{model}{required; a length one character vector.}

\item{temperature}{optional; defaults to \code{1}; a length one numeric vector with the value between \code{0} (More analytical) and \code{2} (More creative).}

\item{top_p}{optional; defaults to \code{1}; a length one numeric vector with the value between \code{0} and \code{1}.}

\item{n}{optional; defaults to \code{1}; a length one numeric vector with the integer value greater than \code{0}.}

\item{max_tokens}{optional; defaults to \verb{(4096 - prompt tokens)}; a length one numeric vector with the integer value greater than \code{0}.}

\item{presence_penalty}{optional; defaults to \code{0}; a length one numeric vector with a value between \code{-2} and \code{2}.}

\item{frequency_penalty}{optional; defaults to \code{0}; a length one numeric vector with a value between \code{-2} and \code{2}.}

\item{progress}{optional; a length one logical vector. Defaults to \code{TRUE}. Determines whether to show a progress bar in the console.}
}
\value{
A dataframe with the output column(s) created
}
\description{
gemini
}
