% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_verify.R
\name{google_verify}
\alias{google_verify}
\title{google_verify}
\usage{
google_verify(...)
}
\arguments{
\item{...}{}
}
\value{
True if the parameters are valid, False otherwise
}
\description{
google_verify
}
