% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{gpt}
\alias{gpt}
\title{gpt}
\usage{
gpt(
  source,
  input,
  output = "output",
  prompt,
  return_invisible = FALSE,
  iterations = 1,
  model = "gpt-3.5-turbo",
  temperature = 1,
  top_p = 1,
  n = 1,
  max_tokens = 4096,
  presence_penalty = 0,
  frequency_penalty = 0,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL,
  progress = FALSE
)
}
\arguments{
\item{source}{required; A source dataframe}

\item{input}{required; A column name in the source dataframe}

\item{output}{optional; A string of a column name (Or a vector of strings) to be created in the source dataframe storing the output of the models. Defaults to \code{output}.}

\item{prompt}{required; A string (Or vector of Strings for handling multiple operations at the same time) of a system message to be sent to the AI model.}

\item{return_invisible}{optional; A boolean to return just the output (TRUE) or an llm object containing model metadata (FALSE). Defaults to \code{FALSE}.}

\item{iterations}{DEV ONLY. SUPPORT COMING SOON. Number of completions to generate. Integer. Defaults to \code{1}.}

\item{model}{required; a length one character vector.}

\item{temperature}{optional; defaults to \code{1}; a length one numeric vector with the value between \code{0} and \code{2}.}

\item{top_p}{optional; defaults to \code{1}; a length one numeric vector with the value between \code{0} and \code{1}.}

\item{n}{optional; defaults to \code{1}; a length one numeric vector with the integer value greater than \code{0}.}

\item{max_tokens}{optional; defaults to \verb{(4096 - prompt tokens)}; a length one numeric vector with the integer value greater than \code{0}.}

\item{presence_penalty}{optional; defaults to \code{0}; a length one numeric vector with a value between \code{-2} and \code{2}.}

\item{frequency_penalty}{optional; defaults to \code{0}; a length one numeric vector with a value between \code{-2} and \code{2}.}

\item{openai_api_key}{required; defaults to \code{Sys.getenv("OPENAI_API_KEY")} (i.e., the value is retrieved from the \code{.Renviron} file); a length one character vector. Specifies OpenAI API key.}

\item{openai_organization}{optional; defaults to \code{NULL}; a length one character vector. Specifies OpenAI organization.}

\item{progress}{optional; a length one logical vector. Defaults to \code{FALSE}. If \code{TRUE}, a progress bar will be displayed.}
}
\value{
A dataframe with the output column(s) created
}
\description{
gpt
}
