% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_model_provider.R
\name{is_valid_model_provider}
\alias{is_valid_model_provider}
\title{Checks if a string is a valid model provider or not}
\usage{
is_valid_model_provider(provider)
}
\arguments{
\item{provider}{required; The string to check if it is a valid model provider or not.}
}
\value{
boolean of if it is a model provider or not.
}
\description{
is_valid_model_provider
}
