% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_llm_completion.R
\name{new_llm_completion}
\alias{new_llm_completion}
\title{new_llm_completion}
\usage{
new_llm_completion(
  x = character(),
  Call = NULL,
  Prompt = NULL,
  Model = NULL,
  Model_Provider = NULL,
  Date = NULL,
  Temperature = NULL,
  Raw = NULL
)
}
\arguments{
\item{x}{The vector to be made into a new llm_completion object}

\item{Prompt}{The prompt used to generate the completion}

\item{Model}{The model used to generate the completion}

\item{Model_Provider}{The provider of the model used to generate the completion}

\item{Date}{The date the completion was generated}

\item{Temperature}{The temperature used to generate the completion}

\item{Raw}{A sample of the raw completion data}
}
\value{
A new llm_completion object
}
\description{
new_llm_completion
}
