% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_verify.R
\name{openai_verify}
\alias{openai_verify}
\title{openai_verify}
\usage{
openai_verify(...)
}
\arguments{
\item{...}{}
}
\value{
True if the parameters are valid, False otherwise.
}
\description{
openai_verify
}
