% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.llm_completion_DBAI.R
\name{print.llm_completion_DBAI}
\alias{print.llm_completion_DBAI}
\title{Print Method for llm_completion objects}
\usage{
\method{print}{llm_completion_DBAI}(x, ...)
}
\arguments{
\item{x}{The \code{llm_completion} object to be printed.}

\item{...}{optional. Other arguments to be passed to the print method.}
}
\value{
invisible
}
\description{
Print Method for llm_completion objects
}
