% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.llm_completion_DBAI.R
\name{summary.llm_completion_DBAI}
\alias{summary.llm_completion_DBAI}
\title{Summary Method for llm_completion objects}
\usage{
\method{summary}{llm_completion_DBAI}(x, ...)
}
\arguments{
\item{x}{The \code{llm_completion} object to be summarized.}

\item{...}{optional. Other arguments to be passed to the print method.}
}
\value{
invisible
}
\description{
Summary Method for llm_completion objects
}
