% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBAI-package.R
\docType{package}
\name{DBAI-package}
\alias{DBAI}
\alias{DBAI-package}
\title{DBAI: AI Tools for R Datasets}
\description{
R Wrapper for AI API's. This package is designed to make it easier to access and use AI API's in R. This package is designed to be used with the OpenAI API, Gemini API, and Claude API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Adam-Hubbs/DBAI}
  \item \url{https://adam-hubbs.github.io/DBAI/}
  \item Report bugs at \url{https://github.com/Adam-Hubbs/DBAI/issues}
}

}
\author{
\strong{Maintainer}: Adam Hubbs \email{amhubbs17@gmail.com}

}
\keyword{internal}
