% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.character.R
\name{gpt.character}
\alias{gpt.character}
\title{gpt}
\usage{
\method{gpt}{character}(
  source,
  prompt = "",
  progress = TRUE,
  repair = FALSE,
  iterations = 1,
  model = "gpt-3.5-turbo",
  temperature = NULL,
  top_p = NULL,
  n = NULL,
  presence_penalty = NULL,
  frequency_penalty = NULL,
  max_tokens = 4096,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL,
  call = rlang::caller_env(),
  output = "output",
  parentInfo = NULL
)
}
\description{
gpt
}
