% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{gpt}
\alias{gpt}
\title{gpt}
\usage{
gpt(
  source,
  input = "input",
  output = "output",
  systemMessage,
  model = "gpt-3.5-turbo",
  temperature = 1,
  top_p = NULL,
  max_tokens = 4096,
  ...
)
}
\arguments{
\item{source}{A source dataframe}

\item{input}{A column name in the source dataframe}

\item{output}{A string of a column name (Or a vetor of strings) to be created in the source dataframe.}

\item{systemMessage}{A string (Or vector of Strings for handling multiple operations at the same time) of a system message to be sent to the GPT model}

\item{model}{The OpenAI Model to be called}

\item{temperature}{Temperature parameter for OpenAI's models}

\item{top_p}{top_p parameter for OpenAI's models}

\item{max_tokens}{Maximum tokens to be generated by OpenAI's models}

\item{...}{}
}
\value{
A dataframe with the output column(s) created
}
\description{
gpt
}
\examples{
sample_df <- gpt(sample_df, input = "demo", output = "Vote", systemMessage = prompt, model = "gpt-3.5-turbo", max_tokens = 5)
}
